+++ expense-service/src/main/java/com/appswella/wisepaise/model/ExpenseGroup.java
@@ -7,6 +7,7 @@ import lombok.NoArgsConstructor;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 
@@ -26,8 +27,8 @@ public class ExpenseGroup {
     private List<String> exGroupMembers;
     private List<Expense> expenses;
     private String exGroupOwnerId;
-    private double exGroupIncome;
-    private double exGroupExpenses;
-    private Map<String, Double> exGroupMembersBalance;
+    private BigDecimal exGroupIncome;
+    private BigDecimal exGroupExpenses;
+    private Map<String, BigDecimal> exGroupMembersBalance;
     private List<Map<String, Object>> exGroupMembersSettlements;
 }
